Minor bug fixes: org.jgrapht.alg.linkprediction SaltonIndexLinkPrediction: L86 double cast org.jgrapht.alg.matching GreedyWeightedMatching: L117,119 double cast org.jgrapht.alg.cycle BergeGraphInspector: L648 double cast org.jgrapht.alg.matching PathGrowingWeightedMatching: L171,243 get() -> orElse(null) org.jgrapht.alg.cycle Cycles: L86 get() -> orElse(null) org.jgrapht.demo HelloJGraphT: L79 get() -> orElse(null) org.jgrapht.graph.guava ImmutableValueGraphAdapter: L220 get() -> orElse(null) org.jgrapht.graph.guava MutableValueGraphAdapter: L334 get() -> orElse(null) org.jgrapht.nio.matrix MatrixExporter: L249 double cast org.jgrapht.opt.graph.sparse.specifics IncomingNoReindexSparseDirectedSpecifics: L73 long cast

+++ jgrapht-core/src/main/java/org/jgrapht/alg/linkprediction/SaltonIndexLinkPrediction.java
@@ -83,7 +83,7 @@ public class SaltonIndexLinkPrediction<V, E>
         Set<V> intersection = new HashSet<>(gu);
         intersection.retainAll(gv);
 
-        return (double) intersection.size() / Math.sqrt(du * dv);
+        return (double) intersection.size() / Math.sqrt((double) (du * dv));
     }
 
 }
