+++ jgrapht-opt/src/main/java/org/jgrapht/opt/graph/sparse/specifics/IncomingNoReindexSparseDirectedSpecifics.java
@@ -70,7 +70,7 @@ public class IncomingNoReindexSparseDirectedSpecifics
         if (inIncidenceMatrix == null) {
             indexIncomingEdges();
         }
-        return outIncidenceMatrix.nonZeros(vertex) + inIncidenceMatrix.nonZeros(vertex);
+        return (long) (outIncidenceMatrix.nonZeros(vertex) + inIncidenceMatrix.nonZeros(vertex));
     }
 
     @Override
